Add EngineConfigField dataclass
Add an EngineConfigField dataclass in common/types.py.
Engines can define any number of these fields in their configuration
dictionary.
The main app can then read the label, type and default value
in order to display the field to the user and to receive inputs.
This allows a uniform way to handle custom engine configurations.