Add common/loader.py and load_classes function
Add common/loader.py utility module and the load_classes function.
The load_classes function will take a module and a base class as
arguments.
It then loads all the submodules in that package,
loads all classes that are subclasses of the base class,
and returns a list of those subclasses.

This is meant to be used by the main script to load all strategies and
engines.